[GLUTEN-7031] Move iterator wrappers to gluten-core
